DatabaseAdministration/Development/ecommerce/cart_service/app/requirements.txt
fastapi
uvicorn
sqlalchemy
databases
python-dotenv
python-jose
pydantic
PyJWT